+++ CHANGELOG.md
@@ -35,7 +35,7 @@ and this project adheres to [Semantic Versioning].
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://gitlab.smoose.nl/development/domain-migrator/tree/master
+[Unreleased]: https://github.com/Smoose-bv/whmcspy
 [0.1.4]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.4
 [0.1.3]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.3
 [0.1.2]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.2