+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning].
 ## [Unreleased]
 ### Added
 - `get_orders` method
+- `cancel_order` method
 
 ## [0.1.5] - 2019-02-05
 ### Added