+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning].
 - `get_orders` method
 - `cancel_order` method
 - `delete_order` method
+- `pending_order` method
 
 ## [0.1.5] - 2019-02-05
 ### Added